requirements.txt
# OpenClaw-Bot Requirements
python-telegram-bot
requests
python-dotenv
